reqested libovr version of the egocentric vr hand tracking sample
github issue #3 request was to provide some reference code for the Rift which uses LibOVR.
This is being checked into a branch at this time as it is not yet a fully debugged and supported sample.
Only tested with VisualStudio 2015, only SR300 tested at this time.
Same limitations apply regarding single right hand only usage.